@@ -38,17 +38,18 @@
     "build": "npm run build:client && npm run build:client-no-css && npm run build:ssr && npm run build:ssr-no-css",
     "watch": "webpack --config ./build/webpack.client.config.js --progress --hide-modules --watch",
     "test:types": "tsc -p types/test",
-    "docs:dev": "vuepress dev docs",
-    "docs:build": "vuepress build docs"
+    "docs:dev": "vitepress dev docs",
+    "docs:build": "vitepress build docs",
+    "docs:serve": "vitepress serve docs"
   },
   "license": "MIT",
   "browserslist": "> 0.25%, not dead",
   "devDependencies": {
     "@babel/core": "^7.2.0",
-    "@babel/preset-env": "^7.2.0",
     "@babel/plugin-proposal-optional-chaining": "^7.10.3",
+    "@babel/preset-env": "^7.2.0",
     "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/test-utils": "1.0.0-beta.31",
+    "@vue/test-utils": "^2.0.0-rc.6",
     "babel-jest": "^23.6.0",
     "babel-loader": "^8.0.4",
     "css-loader": "^1.0.1",
@@ -60,27 +61,30 @@
     "jest": "^23.6.0",
     "jest-serializer-vue": "^2.0.2",
     "mini-css-extract-plugin": "^0.4.5",
-    "node-sass": "^4.13.1",
+    "node-sass": "^4.14.1",
     "optimize-css-assets-webpack-plugin": "^5.0.1",
+    "postcss": "^8.0.0",
     "sass-loader": "^7.1.0",
+    "terser-webpack-plugin": "^4.2.3",
     "ts-loader": "^6.2.2",
     "uglifyjs-webpack-plugin": "^2.2.0",
-    "vue": "2.6.11",
+    "vitepress": "^0.13.0",
+    "vue": "^3",
     "vue-hot-reload-api": "^2.0.8",
     "vue-jest": "^3.0.1",
-    "vue-loader": "^15.0.0",
-    "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "2.6.11",
-    "vue-test-utils": "^1.0.0-beta.11",
-    "vuepress": "^1.5.2",
+    "vue-loader": "^16.2.0",
+    "vue-style-loader": "^4.1.3",
     "webpack": "^4.42.1",
     "webpack-cli": "^3.3.11",
     "webpack-merge": "^4.2.2"
   },
   "peerDependencies": {
-    "vue": "^2.6.11"
+    "vue": "^3"
   },
   "dependencies": {
-    "resize-observer-polyfill": "^1.5.1"
+    "@vue/compiler-sfc": "^3.0.11",
+    "@vue/component-compiler-utils": "^3.2.0",
+    "resize-observer-polyfill": "^1.5.1",
+    "tiny-emitter": "^2.1.0"
   }
 }
